pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "automgr"
version = "0.1.0"
description = "AutoMGR - geração de Mapa de Gerenciamento de Riscos (MGR) via IA"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv>=1.0.0",
  "openai>=1.0.0",
  "google-generativeai>=0.5.0",
  "groq>=0.9.0",
]

[project.scripts]
automgr = "automgr.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

